the full graph is working well, we need to add more nodes to create the final report answer and also to use the retireved data in order for an llm tool to build an answer from it, also we need at the beginning to check if url or document exist in our db, if it exist we go stratight to the retrieval path, if not we follow the full process of embedding. we need to add langfuse also

+++ lib_helpers/embedding_and_retrieval.py
@@ -25,7 +25,7 @@ load_dotenv(dotenv_path='.env', override=False)
 load_dotenv(dotenv_path=".vars", override=True)
 
 # Use Ollama to create embeddings
-embeddings = OllamaEmbeddings(model="mistral:7b", temperature=int(os.getenv("EMBEDDINGS_TEMPERATURE")))
+embeddings = OllamaEmbeddings(model="mistral:7b", temperature=float(os.getenv("EMBEDDINGS_TEMPERATURE")))
 
 # DB CREATION
 def connect_db() -> psycopg2.extensions.connection:
